+++ src/services/configuration.ts
@@ -6,13 +6,8 @@ import * as path from 'path';
 import { fcConnection, ForceService } from '.';
 import * as deepmerge from 'deepmerge'
 
-export const defautlOptions = {
+export const defautlOptions: Config = {
 	apiVersion: vscode.workspace.getConfiguration('force')['defaultApiVersion'],
-	autoRefresh: false,
-	browser: 'Google Chrome Canary',
-	checkForFileChanges: false,
-	debugFilter: 'USER_DEBUG|FATAL_ERROR',
-	debugOnly: true,
 	deployOptions: {
 		'allowMissingFiles': true,
 		'checkOnly': false,
@@ -23,19 +18,11 @@ export const defautlOptions = {
 		'singlePackage': true,
 		'testLevel': 'NoTestRun'
 	},
-	maxFileChangeNotifications: 15,
-	maxQueryHistory: 10,
-	maxQueryResultsPerPage: 250,
-	outputQueriesAsCSV: true,
 	overwritePackageXML: false,
 	poll: 1500,
 	pollTimeout: 1200,
 	prefix: '',
-	revealTestedClass: false,
-	showFilesOnOpen: true,
-	showFilesOnOpenMax: 3,
 	showTestCoverage: true,
-	showTestLog: true,
 	spaDist: '',
 	src: 'src',
 	staticResourceCacheControl: 'Private',
@@ -87,6 +74,21 @@ export function readConfigFile(userName: string, service?: ForceService): Config
 			userName, 'settings.json');
 		if(fs.existsSync(configPath)) {
 			config = fs.readJsonSync(configPath);
+			// these are temporary, just to purge old settings. Will remove in a future release ===
+			delete config['autoRefresh'];
+			delete config['browser'];
+			delete config['checkForFileChanges'];
+			delete config['debugFilter'];
+			delete config['debugOnly'];
+			delete config['maxFileChangeNotifications'];
+			delete config['maxQueryHistory'];
+			delete config['maxQueryResultsPerPage'];
+			delete config['outputQueriesAsCSV'];
+			delete config['revealTestedClass'];
+			delete config['showFilesOnOpen'];
+			delete config['showFilesOnOpenMax'];
+			delete config['showTestLog'];
+			// ====================================================================================
 		} else {
 			config.username = userName;
 		}